+++ src/main/java/insane96mcp/mobspropertiesrandomness/data/json/condition/MPRBlockInCondition.java
@@ -3,17 +3,16 @@ package insane96mcp.mobspropertiesrandomness.data.json.condition;
 import com.google.gson.*;
 import com.google.gson.annotations.JsonAdapter;
 import insane96mcp.insanelib.data.IdTagMatcher;
-import net.minecraft.core.BlockPos;
 import net.minecraft.world.entity.LivingEntity;
 
 import java.lang.reflect.Type;
 import java.util.List;
 
-@JsonAdapter(MPRBlockOnCondition.Serializer.class)
-public class MPRBlockOnCondition extends MPRCondition {
+@JsonAdapter(MPRBlockInCondition.Serializer.class)
+public class MPRBlockInCondition extends MPRCondition {
     List<IdTagMatcher> blocks;
 
-    public MPRBlockOnCondition(List<IdTagMatcher> blocks, boolean inverted) {
+    public MPRBlockInCondition(List<IdTagMatcher> blocks, boolean inverted) {
         super(inverted);
         this.blocks = blocks;
     }
@@ -21,27 +20,25 @@ public class MPRBlockOnCondition extends MPRCondition {
     @Override
     protected boolean conditionCheck(LivingEntity living) {
         for (IdTagMatcher block : this.blocks) {
-            for (BlockPos pos : BlockPos.betweenClosedStream(living.getBoundingBox()).toList()) {
-                if (block.matchesBlock(living.level().getBlockState(pos)))
-                    return true;
-            }
+            if (block.matchesBlock(living.getBlockStateOn()))
+                return true;
         }
         return false;
     }
 
-    public static class Serializer implements JsonDeserializer<MPRBlockOnCondition>, JsonSerializer<MPRBlockOnCondition> {
+    public static class Serializer implements JsonDeserializer<MPRBlockInCondition>, JsonSerializer<MPRBlockInCondition> {
         @Override
-        public MPRBlockOnCondition deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
+        public MPRBlockInCondition deserialize(JsonElement json, Type typeOfT, JsonDeserializationContext context) throws JsonParseException {
             JsonObject jObject = json.getAsJsonObject();
             JsonArray aBlocks = jObject.getAsJsonArray("blocks");
             if (aBlocks == null)
                 throw new JsonParseException("Missing blocks array");
             List<IdTagMatcher> values = context.deserialize(aBlocks, IdTagMatcher.LIST_TYPE);
-            return new MPRBlockOnCondition(values, MPRCondition.deserializeInverted(jObject));
+            return new MPRBlockInCondition(values, MPRCondition.deserializeInverted(jObject));
         }
 
         @Override
-        public JsonElement serialize(MPRBlockOnCondition src, Type typeOfSrc, JsonSerializationContext context) {
+        public JsonElement serialize(MPRBlockInCondition src, Type typeOfSrc, JsonSerializationContext context) {
             JsonObject jObject = new JsonObject();
             jObject.add("blocks", context.serialize(src.blocks));
             return src.endSerialization(jObject);